<!-- Thank you for your contribution! Please review https://github.com/ray-project/ray/blob/master/CONTRIBUTING.rst before opening a pull request. -->

<!-- Please add a reviewer to the assignee section when you create a PR. If you don't have the access to it, we will shortly find a reviewer and assign them to your PR. -->

## Why are these changes needed?

We found the performance of LSTMs in Rllib to be extremely slow compared to other methods, with a single training iteration of PPO taking 179 seconds (compared to ~9 seconds with a similarly-sized MLP network). This made RNNs/LSTMs, as well as some transformer implementations, completely unusable for our purposes.

However, when profiling, we found this was primarily due to a very slow copy operation:

<img width="363" alt="image" src="https://github.com/ray-project/ray/assets/2100334/9b518388-91ec-4f17-ac2a-d4e90d109a60">

Further investigation revealed that most of this runtime was spent copying the `infos` dict. We determined that the root cause was inconsistent handling of the dictionary in `rnn_sequencing`. While the non-recurrent implementation stores the list of dictionaries as a NumPy array of objects, `rnn_sequencing` instead stores it as a Python list:

<img width="538" alt="image" src="https://github.com/ray-project/ray/assets/2100334/571ce0dc-65e6-43c2-bee5-f04325542894">

We applied a one-line fix to make this behavior consistent and store the list as NumPy array:

```python
# old and slow
f_pad = [None] * length
# new and fast
f_pad = np.full([length], None, dtype=f.dtype)
```

This causes the `copy` function to perform a shallow copy, drastically improving performance by ~6x to around 29 seconds:

<img width="322" alt="image" src="https://github.com/ray-project/ray/assets/2100334/e44c50f5-b7a2-42fb-9483-c219cd65ab81">

However, we found that the training loop was still spending a lot of time in rnn_sequencing. We traced this down to a slow element-wise copy into an array. We instead replaced this with a vectorized copy:

```python
# old and slow
for seq_offset in range(len_):
    f_pad[seq_base + seq_offset] = f[i]
    i += 1
# new and fast
f_pad[seq_base : seq_base + len_] = f[i : i + len_]
i += len_
```

This further improved performance (in this sample, we also removed the 1-time summary logging, but did not remove this in this PR):

<img width="360" alt="image" src="https://github.com/ray-project/ray/assets/2100334/378b2a7e-a055-4168-9383-5cb52cabc185">

Combined, these changes reduced the run time of the training step from 179 seconds to 15 seconds, approximately a 12x speedup and competitive with training an MLP.

## Related issue number

<!-- For example: "Closes #1234" -->

## Checks

- [ ] I've signed off every commit(by using the -s flag, i.e., `git commit -s`) in this PR.
- [ ] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for https://docs.ray.io/en/master/.
    - [ ] I've added any new APIs to the API Reference. For example, if I added a 
           method in Tune, I've added it in `doc/source/tune/api/` under the 
           corresponding `.rst` file.
- [ ] I've made sure the tests are passing. Note that there might be a few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [ ] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(